adds grade_criteria
